+++ app/models/location.rb
@@ -2,7 +2,7 @@ class Location < ActiveRecord::Base
   belongs_to :library
 
   UAL_SHIELD_LIBRARIES = [
-    Location.find_by(short_code: :uainternet)&.name,
-    Location.find_by(short_code: :neosfree)&.name
+    Location.find_by(short_code: 'UAINTERNET')&.name,
+    Location.find_by(short_code: 'NEOS_FREE')&.name
   ].freeze
 end