+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20200128205653) do
+ActiveRecord::Schema.define(version: 20200128220027) do
 
   create_table "backup_libraries", force: :cascade do |t|
     t.string   "short_code",     limit: 255
@@ -26,6 +26,7 @@ ActiveRecord::Schema.define(version: 20200128205653) do
 
   create_table "backup_locations", force: :cascade do |t|
     t.string   "short_code",        limit: 255
+    t.string   "old_short_code",    limit: 255
     t.string   "name",              limit: 255
     t.string   "url",               limit: 255
     t.integer  "backup_library_id", limit: 4
@@ -73,12 +74,13 @@ ActiveRecord::Schema.define(version: 20200128205653) do
   end
 
   create_table "locations", force: :cascade do |t|
-    t.string   "short_code", limit: 255
-    t.string   "name",       limit: 255
-    t.string   "url",        limit: 255
-    t.integer  "library_id", limit: 4
-    t.datetime "created_at",             null: false
-    t.datetime "updated_at",             null: false
+    t.string   "short_code",     limit: 255
+    t.string   "name",           limit: 255
+    t.string   "url",            limit: 255
+    t.integer  "library_id",     limit: 4
+    t.datetime "created_at",                 null: false
+    t.datetime "updated_at",                 null: false
+    t.string   "old_short_code", limit: 255
   end
 
   add_index "locations", ["library_id"], name: "index_locations_on_library_id", using: :btree