+++ spec/models/backup_location_spec.rb
@@ -4,6 +4,7 @@ RSpec.describe BackupLocation, type: :model do
   subject do
     described_class.new(
       short_code: 'UASCITECH',
+      old_short_code: 'uascitech',
       name: 'University of Alberta Cameron - Science & Technology',
       url: 'https://www.library.ualberta.ca/locations/cameron'
     )
@@ -11,6 +12,7 @@ RSpec.describe BackupLocation, type: :model do
 
   describe 'Properties' do
     it { is_expected.to have_attributes(short_code: 'UASCITECH') }
+    it { is_expected.to have_attributes(old_short_code: 'uascitech') }
   end
 
   describe 'Associations' do