+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and releases in Discovery project adheres to [Semantic Versioning](http://semver
 ### Removed
 - update tests to address deprecations [PR#1901](https://github.com/ualbertalib/discovery/pull/1901)
 
+### Changed
+- data sanity regression test values for databases [PR#]()
+
 ## [3.0.116] - 2020-02-02
 ### Added
 - Added "See License Terms" link to SFX e-journal record holdings. [#1893]